+++ src/main/java/org/fugerit/java/sql/safe/helper/SqlExceptionHelper.java
@@ -1,0 +1,19 @@
+package org.fugerit.java.sql.safe.helper;
+
+import java.sql.SQLException;
+
+public class SqlExceptionHelper {
+
+    public SqlExceptionHelper() {}
+
+    public static SQLException findSQLException(Throwable e ) {
+        if ( e == null ) {
+            return null;
+        } else if ( e instanceof SQLException) {
+            return (SQLException) e;
+        } else {
+            return findSQLException( e.getCause() );
+        }
+    }
+
+}