+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- SqlExceptionHelper utility, findSQLException()
+
 ### Changed
 
 - fj-version set to 8.6.5